Frontend: Add common error page
In case of an error there can either be a redirect made to '/error' or the error page rendered in place. The former is useful for the auth service which redirects to the error page in case of login/logout failure. And the latter is used by the frontend itself.

#66