Frontend: Communicate import errors to the user
Instead of locally storing which repos are being imported, and querying the scheduler about the progress, we now look at memcached. There the scheduler now distributes the state of running and finished tasks. Thanks to memcached lifetimes, we can show errors to the user in the fronted, even after tasks have finished, without having to gc/evict stale information ourselves.