API-Bridge: Do not assume repo URL structure
GitLab URLs consist of a base URL and a project ID. The base URL consists of the domain and maybe a bit of the path when GitLab is hosted on a relative path behind a proxy. The project ID on the other hand can be prefixed with (sub-)groups. Hence we cannot assume there is only one group as we did previously and instead figure out where each part of the provided URL starts/ends.