Docker: Only use a single busybox binary in the fluentbit image
Instead of copying the busybox binary for each shell command, we now copy the binary once and then point symlinks to it, the way busybox is intended to be used. The installation is divided into two steps: First we setup a shell and then run a shell script that handles the rest of the installation. We also now install to `/bin` instead of `/usr/local/bin`.