Nginx: Resolve DNS during runtime
By default nginx does DNS lookups of the proxied upstream service once at startup. It also crashes when an upstream is not resolvable yet. Whenever a service restarts on a different IP nginx cannot find it anymore and would itself need to be restarted (or reconfigured via nginx -s reload). By setting a resolver and custom TTL in the config, nginx resolves the DNS names also during runtime. However, nginx uses its own DNS resolver that needs a bit of handholding as it does not look at the `/etc/resolv.conf`. Hence, we need to set the DNS server host and append the DNS search part ourselves. Furthermore, we need to use a variable in the proxy_pass command to delay evaluation, which then disables automatic path rewriting. Therefore, we have to use the rewrite command to remove the location path prefix ourselves. The DNS host get read from `/etc/resolv.conf` as part of a custom startup script. For this reason we call the entrypoint of the nginx docker image via exec now.