Scheduler: Prevent callback race condition
Previously a task sent a fetch request to the next service for a snapshot and only then started waiting for callbacks. In the time until we started accepting callbacks, any service fast enough to respond could kill the task. On the kubernetes cluster somehow the scheduler often inexplicably hangs just before we start waiting, which triggers the race condition.
Now we use a queue based system, where callbacks get always accepted to be queued, and then get processed when we are ready.